+++ storybook/stories/staking/Staking.stories.js
@@ -26,7 +26,7 @@ import {
 } from './Undelegate.stories';
 
 const defaultPercentage = 10;
-const defaultStartDateTime = new Date('2019-09-26');
+const defaultStartDateTime = new Date('2019-12-31');
 const startDateTimeKnob = (name, defaultValue) => {
   const stringTimestamp = date(name, defaultValue);